feat(patient): implement update patient profile use case
- update PatientServiceImpl to support patient profile updates
- include restriction resolution using restriction codes
- integrate physiological conditions update
- add validation and exception handling with ResourceNotFoundException
- map restrictions to DTOs using RestrictionMapper
- update user and patient profile fields in a single transactional operation"

+++ nutrition/src/main/java/com/lirium/nutrition/dto/request/PatientProfileCreateRequestDTO.java
@@ -20,7 +20,7 @@ public record PatientProfileCreateRequestDTO(
 
         @DecimalMin("20.0")
         @DecimalMax("400.0")
-        BigDecimal weight,
+        Integer weight,
 
         @Min(80)
         @Max(250)
