+++ nutrition/src/main/java/com/lirium/nutrition/dto/response/PatientProfileHistorySummaryDTO.java
@@ -8,6 +8,6 @@ public record PatientProfileHistorySummaryDTO(
         Long id,
         Long patientProfileId,
         LocalDate visitDate,
-        BigDecimal weight
+        Integer weight
 
 ) {}
